refactor(cashu): extract framework-free Cashu protocol primitives into @agicash/cashu
First slice of lib-ifying ~/lib/cashu: move the pure, framework-free protocol files into a
new @agicash/cashu leaf package so the SDK cashu domain can consume them without dragging
React in.

- error-codes, types (ProofSchema), secret, proof, token, payment-request,
  protocol-extensions, blind-signature-matching (+ their 5 tests) -> packages/cashu/src
  (byte-identical git mv; only types.ts/secret.ts rewired to @agicash/utils for zod/json)
- the @agicash/cashu barrel deliberately exports NO React (the lib's framework-free contract)
- old ~/lib/cashu/<file> paths kept as transitional re-export shims so the web barrel and the
  ~20 consumers stay untouched
- @cashu/cashu-ts, @noble/hashes, @noble/curves hoisted to the root catalog (now shared by web
  + @agicash/cashu); cashu-ts pinned so the lib and web can't drift on the subclass overrides
- @agicash/cashu deps: @agicash/utils, @cashu/cashu-ts, @noble/hashes, zod

Phase 0. typecheck, biome, build all green; tests preserved (133 total: web 84, utils 14,
cashu 35); framework-free gate verified. Still in ~/lib/cashu (Chunk C): the utils.ts/wallet
split, mint-validation, the 2 subscription managers; staying in web: the React hook +
animated-qr.
