+++ apps/web-wallet/package.json
@@ -23,10 +23,10 @@
     "@agicash/opensecret": "catalog:",
     "@agicash/qr-scanner": "0.1.2",
     "@agicash/utils": "workspace:*",
-    "@cashu/cashu-ts": "3.6.1",
+    "@cashu/cashu-ts": "catalog:",
     "@noble/ciphers": "1.3.0",
-    "@noble/curves": "1.9.7",
-    "@noble/hashes": "1.8.0",
+    "@noble/curves": "catalog:",
+    "@noble/hashes": "catalog:",
     "@radix-ui/react-checkbox": "1.3.3",
     "@radix-ui/react-dialog": "1.1.6",
     "@radix-ui/react-dropdown-menu": "2.1.15",
