+++ package.json
@@ -6,6 +6,9 @@
     "packages": ["apps/*", "packages/*"],
     "catalog": {
       "@agicash/opensecret": "0.1.0",
+      "@cashu/cashu-ts": "3.6.1",
+      "@noble/curves": "1.9.7",
+      "@noble/hashes": "1.8.0",
       "@stablelib/base64": "2.0.1",
       "@stablelib/chacha20poly1305": "2.0.1",
       "@supabase/supabase-js": "2.95.2",
