+++ packages/cashu/src/index.ts
@@ -1,0 +1,10 @@
+// Framework-free Cashu protocol lib. MUST NOT export React/UI (hooks, components) —
+// the SDK domain compiles against this barrel and relies on it being framework-free.
+export * from './error-codes';
+export * from './types';
+export * from './secret';
+export * from './proof';
+export * from './token';
+export * from './payment-request';
+export * from './protocol-extensions';
+export * from './blind-signature-matching';
