+++ packages/cashu/src/payment-request.test.ts
