+++ packages/cashu/src/proof.test.ts
