+++ packages/cashu/src/secret.test.ts
