+++ packages/cashu/src/token.test.ts
